requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-debug-toolbar==1.9.1
dlib==19.7.0
idna==2.6
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
Pillow==4.3.0
pytz==2017.3
requests==2.18.4
sqlparse==0.2.4
urllib3==1.22
